rename editor
